# docs/swagger.yaml
definitions:
  dto.CartItemRequest:
    properties:
      product_id:
        type: integer
      quantity:
        type: integer
    type: object
  dto.CartItemResponse:
    properties:
      cart_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      product_id:
        type: integer
      quantity:
        type: integer
    type: object
  dto.CartItemUpdateRequest:
    properties:
      quantity:
        type: integer
    type: object
...
info:
  contact: {}
  description: A Shop wrote by Go using Gin framework
  title: Shop
  version: "1.0"
paths:
...
  /api/v1/cart/item:
    delete:
      consumes:
      - application/json
      description: Deletes all items from cart
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Deletes all items from cart
      tags:
      - Cart
    get:
      consumes:
      - application/json
      description: Gets all items in cart
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.CartItemResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Gets all items in cart
      tags:
      - Cart
    post:
      consumes:
      - application/json
      description: Add existing product to cart
      parameters:
      - description: Data for add item to cart
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/dto.CartItemRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CartItemResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Add product to cart
      tags:
      - Cart
  /api/v1/cart/item/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes item from cart
      parameters:
      - description: Cart item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Deletes item from cart
      tags:
      - Cart
    patch:
      consumes:
      - application/json
      description: Updates cart item quantity
      parameters:
      - description: Cart item ID
        in: path
        name: id
        required: true
        type: integer
      - description: Item quantity
        in: body
        name: Quantity
        required: true
        schema:
          $ref: '#/definitions/dto.CartItemUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Updates cart item quantity
      tags:
      - Cart
...
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"